Add files via upload
an .ipynb file for analyzing the electrostatic interaction energy data generated for the Coarse-Grained MD simulations containing one Hfq protein (with or without polyphosphate and DNA). One should download the attached sub-folders containing the energy files, then insert their paths to the required locations in the .ipynb file and run the file to produce the desired plots.